+++ msl_motion/src/settings.cpp
@@ -99,7 +99,7 @@ void settings_init() {
     }*/
 
 
-    default_settings.wheelRadius = motion->get<int>("Motion","MotionControl","WheelRadius",NULL);
+    default_settings.wheelRadius = motion->get<double>("Motion","MotionControl","WheelRadius",NULL);
     default_settings.robotRadius = motion->get<double>("Motion","MotionControl","RobotRadius",NULL);
